#18289: Move yolov8s_world model from experimental to demos folder
