+++ models/demos/yolov8s_world/demo/images/elephants.jpg
