+++ models/demos/yolov8s_world/runner/performant_runner_infra.py
@@ -6,15 +6,14 @@
 import torch
 import torch.nn.functional as F
 from loguru import logger
+from ttnn.model_preprocessing import preprocess_model_parameters
 
 import ttnn
-from models.experimental.yolov8s_world.demo.demo_utils import load_torch_model
-
-from models.experimental.yolov8s_world.tt.ttnn_yolov8s_world import TtYOLOWorld
+from models.demos.yolov8s_world.demo.demo_utils import load_torch_model
+from models.demos.yolov8s_world.tt.ttnn_yolov8s_world import TtYOLOWorld
+from models.demos.yolov8s_world.tt.ttnn_yolov8s_world_utils import create_custom_preprocessor, move_to_device
 from models.utility_functions import divup, is_wormhole_b0
 from tests.ttnn.utils_for_testing import assert_with_pcc
-from ttnn.model_preprocessing import preprocess_model_parameters
-from models.experimental.yolov8s_world.tt.ttnn_yolov8s_world_utils import create_custom_preprocessor, move_to_device
 
 
 class YOLOv8sWorldPerformanceRunnerInfra:
